Migrate HTTP handler and client into http subpackage.
