+++ fragment.go
@@ -1782,8 +1782,7 @@ func (s *FragmentSyncer) SyncFragment() error {
 		}
 
 		// Retrieve remote blocks.
-		client := NewInternalHTTPClientFromURI(&node.URI, s.RemoteClient)
-		blocks, err := client.FragmentBlocks(context.Background(), s.Fragment.Index(), s.Fragment.Field(), s.Fragment.Slice())
+		blocks, err := s.Cluster.InternalClient.FragmentBlocks(context.Background(), nil, s.Fragment.Index(), s.Fragment.Field(), s.Fragment.Slice())
 		if err != nil && err != ErrFragmentNotFound {
 			return errors.Wrap(err, "getting blocks")
 		}
@@ -1847,7 +1846,7 @@ func (s *FragmentSyncer) syncBlock(id int) error {
 
 	// Read pairs from each remote block.
 	var pairSets []PairSet
-	var clients []InternalClient
+	var uris []*URI
 	for _, node := range s.Cluster.SliceNodes(f.Index(), f.Slice()) {
 		if s.Node.ID == node.ID {
 			continue
@@ -1858,11 +1857,11 @@ func (s *FragmentSyncer) syncBlock(id int) error {
 			return nil
 		}
 
-		client := NewInternalHTTPClientFromURI(&node.URI, s.RemoteClient)
-		clients = append(clients, client)
+		uri := &node.URI
+		uris = append(uris, uri)
 
 		// Only sync the standard block.
-		rowIDs, columnIDs, err := client.BlockData(context.Background(), f.Index(), f.Field(), f.Slice(), id)
+		rowIDs, columnIDs, err := s.Cluster.InternalClient.BlockData(context.Background(), &node.URI, f.Index(), f.Field(), f.Slice(), id)
 		if err != nil {
 			return errors.Wrap(err, "getting block")
 		}
@@ -1885,7 +1884,7 @@ func (s *FragmentSyncer) syncBlock(id int) error {
 	}
 
 	// Write updates to remote blocks.
-	for i := 0; i < len(clients); i++ {
+	for i := 0; i < len(uris); i++ {
 		set, clear := sets[i], clears[i]
 		count := 0
 
@@ -1924,7 +1923,7 @@ func (s *FragmentSyncer) syncBlock(id int) error {
 				Query:  buffers[k].String(),
 				Remote: true,
 			}
-			_, err := clients[i].Query(context.Background(), f.Index(), queryRequest)
+			_, err := s.Cluster.InternalClient.QueryNode(context.Background(), uris[i], f.Index(), queryRequest)
 			if err != nil {
 				return errors.Wrap(err, "executing")
 			}
