+++ http/handler_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package pilosa_test
+package http_test
 
 import (
 	"bytes"
@@ -21,7 +21,7 @@ import (
 	"fmt"
 	"io"
 	"io/ioutil"
-	"net/http"
+	gohttp "net/http"
 	"net/http/httptest"
 	"reflect"
 	"strings"
@@ -29,6 +29,7 @@ import (
 
 	"github.com/gogo/protobuf/proto"
 	"github.com/pilosa/pilosa"
+	"github.com/pilosa/pilosa/http"
 	"github.com/pilosa/pilosa/internal"
 	"github.com/pilosa/pilosa/pql"
 	"github.com/pilosa/pilosa/test"
@@ -49,7 +50,7 @@ func TestHandlerPanics(t *testing.T) {
 	if !bytes.Contains(bufbytes, []byte("PANIC: runtime error: invalid memory address or nil pointer dereference")) {
 		t.Fatalf("expected panic in log, but got: %s", bufbytes)
 	}
-	if w.Code != http.StatusInternalServerError {
+	if w.Code != gohttp.StatusInternalServerError {
 		t.Fatalf("expected internal server error, but got: %v", w.Code)
 	}
 	bodyBytes := w.Body.Bytes()
@@ -69,7 +70,7 @@ func TestHandler_NotFound(t *testing.T) {
 
 	w := httptest.NewRecorder()
 	h.ServeHTTP(w, test.MustNewHTTPRequest("GET", "/no_such_path", nil))
-	if w.Code != http.StatusNotFound {
+	if w.Code != gohttp.StatusNotFound {
 		t.Fatalf("invalid status: %d", w.Code)
 	}
 }
@@ -101,7 +102,7 @@ func TestHandler_Schema(t *testing.T) {
 	h.API.Cluster = test.NewCluster(1)
 	w := httptest.NewRecorder()
 	h.ServeHTTP(w, test.MustNewHTTPRequest("GET", "/schema", nil))
-	if w.Code != http.StatusOK {
+	if w.Code != gohttp.StatusOK {
 		t.Fatalf("unexpected status code: %d", w.Code)
 	} else if body := w.Body.String(); body != `{"indexes":[{"name":"i0","fields":[{"name":"f0"},{"name":"f1","views":[{"name":"standard"}]}]},{"name":"i1","fields":[{"name":"f0","views":[{"name":"standard"}]}]}]}`+"\n" {
 	} else if body := w.Body.String(); body != `{"indexes":[{"name":"i0","fields":[{"name":"f0","options":{"cacheType":"ranked","cacheSize":50000}},{"name":"f1","options":{"cacheType":"ranked","cacheSize":50000},"views":[{"name":"standard"}]}]},{"name":"i1","fields":[{"name":"f0","options":{"cacheType":"ranked","cacheSize":50000},"views":[{"name":"standard"}]}]}]}`+"\n" {
@@ -142,7 +143,7 @@ func TestHandler_Status(t *testing.T) {
 
 	w := httptest.NewRecorder()
 	h.ServeHTTP(w, test.MustNewHTTPRequest("GET", "/status", nil))
-	if w.Code != http.StatusOK {
+	if w.Code != gohttp.StatusOK {
 		t.Fatalf("unexpected status code: %d", w.Code)
 	} else if body := w.Body.String(); body != `{"state":"NORMAL","nodes":[{"id":"node0","uri":{"scheme":"http","host":"host0"},"isCoordinator":false}],"localID":"node0"}`+"\n" {
 		t.Fatalf("unexpected body: %s", body)
@@ -156,9 +157,9 @@ func TestHandler_Info(t *testing.T) {
 
 	w := httptest.NewRecorder()
 	h.ServeHTTP(w, test.MustNewHTTPRequest("GET", "/info", nil))
-	if w.Code != http.StatusOK {
+	if w.Code != gohttp.StatusOK {
 		t.Fatalf("unexpected status code: %d", w.Code)
-	} else if body := w.Body.String(); body != fmt.Sprintf("{\"sliceWidth\":%d}\n", SliceWidth) {
+	} else if body := w.Body.String(); body != fmt.Sprintf("{\"sliceWidth\":%d}\n", pilosa.SliceWidth) {
 		t.Fatalf("unexpected body: %s", body)
 	}
 }
@@ -173,7 +174,7 @@ func TestHandler_ClusterResizeAbort(t *testing.T) {
 
 		w := httptest.NewRecorder()
 		h.ServeHTTP(w, test.MustNewHTTPRequest("POST", "/cluster/resize/abort", nil))
-		if w.Code != http.StatusOK {
+		if w.Code != gohttp.StatusOK {
 			bod, err := ioutil.ReadAll(w.Body)
 			t.Fatalf("unexpected status code: %d, bod: %s, readerr: %v", w.Code, bod, err)
 		} else if body := w.Body.String(); body != `{"info":"complete current job: no resize job currently running"}`+"\n" {
@@ -188,20 +189,20 @@ func TestHandler_MaxSlices(t *testing.T) {
 	hldr := test.MustOpenHolder()
 	defer hldr.Close()
 
-	hldr.MustCreateFragmentIfNotExists("i0", "f0", pilosa.ViewStandard, 1).MustSetBits(30, (1*SliceWidth)+1)
-	hldr.MustCreateFragmentIfNotExists("i0", "f0", pilosa.ViewStandard, 1).MustSetBits(30, (1*SliceWidth)+2)
-	hldr.MustCreateFragmentIfNotExists("i0", "f0", pilosa.ViewStandard, 3).MustSetBits(30, (3*SliceWidth)+4)
+	hldr.MustCreateFragmentIfNotExists("i0", "f0", pilosa.ViewStandard, 1).MustSetBits(30, (1*pilosa.SliceWidth)+1)
+	hldr.MustCreateFragmentIfNotExists("i0", "f0", pilosa.ViewStandard, 1).MustSetBits(30, (1*pilosa.SliceWidth)+2)
+	hldr.MustCreateFragmentIfNotExists("i0", "f0", pilosa.ViewStandard, 3).MustSetBits(30, (3*pilosa.SliceWidth)+4)
 
-	hldr.MustCreateFragmentIfNotExists("i1", "f1", pilosa.ViewStandard, 0).MustSetBits(40, (0*SliceWidth)+1)
-	hldr.MustCreateFragmentIfNotExists("i1", "f1", pilosa.ViewStandard, 0).MustSetBits(40, (0*SliceWidth)+2)
-	hldr.MustCreateFragmentIfNotExists("i1", "f1", pilosa.ViewStandard, 0).MustSetBits(40, (0*SliceWidth)+8)
+	hldr.MustCreateFragmentIfNotExists("i1", "f1", pilosa.ViewStandard, 0).MustSetBits(40, (0*pilosa.SliceWidth)+1)
+	hldr.MustCreateFragmentIfNotExists("i1", "f1", pilosa.ViewStandard, 0).MustSetBits(40, (0*pilosa.SliceWidth)+2)
+	hldr.MustCreateFragmentIfNotExists("i1", "f1", pilosa.ViewStandard, 0).MustSetBits(40, (0*pilosa.SliceWidth)+8)
 
 	h := test.MustNewHandler()
 	h.API.Holder = hldr.Holder
 	h.API.Cluster = test.NewCluster(1)
 	w := httptest.NewRecorder()
 	h.ServeHTTP(w, test.MustNewHTTPRequest("GET", "/slices/max", nil))
-	if w.Code != http.StatusOK {
+	if w.Code != gohttp.StatusOK {
 		t.Fatalf("unexpected status code: %d", w.Code)
 	} else if body := w.Body.String(); body != `{"standard":{"i0":3,"i1":0}}`+"\n" {
 		t.Fatalf("unexpected body: %s", body)
@@ -229,7 +230,7 @@ func TestHandler_Query_Args_URL(t *testing.T) {
 
 	w := httptest.NewRecorder()
 	h.ServeHTTP(w, test.MustNewHTTPRequest("POST", "/index/idx0/query?slices=0,1", strings.NewReader("Count( Bitmap( id=100))")))
-	if w.Code != http.StatusOK {
+	if w.Code != gohttp.StatusOK {
 		t.Fatalf("unexpected status code: %d %s", w.Code, w.Body.String())
 	} else if body := w.Body.String(); body != `{"results":[100]}`+"\n" {
 		t.Fatalf("unexpected body: %q", body)
@@ -270,7 +271,7 @@ func TestHandler_Query_Args_Protobuf(t *testing.T) {
 
 	w := httptest.NewRecorder()
 	h.ServeHTTP(w, req)
-	if w.Code != http.StatusOK {
+	if w.Code != gohttp.StatusOK {
 		t.Fatalf("unexpected status code: %d", w.Code)
 	}
 }
@@ -286,7 +287,7 @@ func TestHandler_Query_Args_Err(t *testing.T) {
 	h.API.Holder = hldr.Holder
 
 	h.ServeHTTP(w, test.MustNewHTTPRequest("POST", "/index/idx0/query?slices=a,b", strings.NewReader("Bitmap(id=100)")))
-	if w.Code != http.StatusBadRequest {
+	if w.Code != gohttp.StatusBadRequest {
 		t.Fatalf("unexpected status code: %d", w.Code)
 	} else if body := w.Body.String(); body != `{"error":"invalid slice argument"}`+"\n" {
 		t.Fatalf("unexpected body: %q", body)
@@ -295,7 +296,7 @@ func TestHandler_Query_Args_Err(t *testing.T) {
 func TestHandler_Query_Params_Err(t *testing.T) {
 	w := httptest.NewRecorder()
 	test.MustNewHandler().ServeHTTP(w, test.MustNewHTTPRequest("POST", "/index/idx0/query?slices=0,1&db=sample", strings.NewReader("Bitmap(id=100)")))
-	if w.Code != http.StatusBadRequest {
+	if w.Code != gohttp.StatusBadRequest {
 		t.Fatalf("unexpected status code: %d", w.Code)
 	} else if body := w.Body.String(); body != `{"error":"db is not a valid argument"}`+"\n" {
 		t.Fatalf("unexpected body: %q", body)
@@ -317,7 +318,7 @@ func TestHandler_Query_Uint64_JSON(t *testing.T) {
 
 	w := httptest.NewRecorder()
 	h.ServeHTTP(w, test.MustNewHTTPRequest("POST", "/index/idx0/query?slices=0,1", strings.NewReader("Count( Bitmap( id=100))")))
-	if w.Code != http.StatusOK {
+	if w.Code != gohttp.StatusOK {
 		t.Fatalf("unexpected status code: %d", w.Code)
 	} else if body := w.Body.String(); body != `{"results":[100]}`+"\n" {
 		t.Fatalf("unexpected body: %q", body)
@@ -340,14 +341,14 @@ func TestHandler_Query_Uint64_Protobuf(t *testing.T) {
 	r := test.MustNewHTTPRequest("POST", "/index/i/query", strings.NewReader("Count(Bitmap(id=100))"))
 	r.Header.Set("Accept", "application/x-protobuf")
 	h.ServeHTTP(w, r)
-	if w.Code != http.StatusOK {
+	if w.Code != gohttp.StatusOK {
 		t.Fatalf("unexpected status code: %d", w.Code)
 	}
 
 	var resp internal.QueryResponse
 	if err := proto.Unmarshal(w.Body.Bytes(), &resp); err != nil {
 		t.Fatal(err)
-	} else if rt := resp.Results[0].Type; rt != pilosa.QueryResultTypeUint64 {
+	} else if rt := resp.Results[0].Type; rt != http.QueryResultTypeUint64 {
 		t.Fatalf("unexpected response type: %d", resp.Results[0].Type)
 	} else if n := resp.Results[0].N; n != 100 {
 		t.Fatalf("unexpected n: %d", n)
@@ -370,7 +371,7 @@ func TestHandler_Query_Bitmap_JSON(t *testing.T) {
 
 	w := httptest.NewRecorder()
 	h.ServeHTTP(w, test.MustNewHTTPRequest("POST", "/index/i/query", strings.NewReader("Bitmap(id=100)")))
-	if w.Code != http.StatusOK {
+	if w.Code != gohttp.StatusOK {
 		t.Fatalf("unexpected status code: %d", w.Code)
 	} else if body := w.Body.String(); body != `{"results":[{"attrs":{"a":"b","c":1,"d":true},"columns":[1,3,66,1048577]}]}`+"\n" {
 		t.Fatalf("unexpected body: %s", body)
@@ -403,7 +404,7 @@ func TestHandler_Query_Row_ColumnAttrs_JSON(t *testing.T) {
 
 	w := httptest.NewRecorder()
 	h.ServeHTTP(w, test.MustNewHTTPRequest("POST", "/index/i/query?columnAttrs=true", strings.NewReader("Bitmap(id=100)")))
-	if w.Code != http.StatusOK {
+	if w.Code != gohttp.StatusOK {
 		t.Fatalf("unexpected status code: %d", w.Code)
 	} else if body := w.Body.String(); body != `{"results":[{"attrs":{"a":"b","c":1,"d":true},"columns":[1,3,66,1048577]}],"columnAttrs":[{"id":3,"attrs":{"x":"y"}},{"id":66,"attrs":{"y":123,"z":false}}]}`+"\n" {
 		t.Fatalf("unexpected body: %s", body)
@@ -428,16 +429,16 @@ func TestHandler_Query_Row_Protobuf(t *testing.T) {
 	r := test.MustNewHTTPRequest("POST", "/index/i/query", strings.NewReader("Bitmap(id=100)"))
 	r.Header.Set("Accept", "application/x-protobuf")
 	h.ServeHTTP(w, r)
-	if w.Code != http.StatusOK {
+	if w.Code != gohttp.StatusOK {
 		t.Fatalf("unexpected status code: %d", w.Code)
 	}
 
 	var resp internal.QueryResponse
 	if err := proto.Unmarshal(w.Body.Bytes(), &resp); err != nil {
 		t.Fatal(err)
-	} else if rt := resp.Results[0].Type; rt != pilosa.QueryResultTypeRow {
+	} else if rt := resp.Results[0].Type; rt != http.QueryResultTypeRow {
 		t.Fatalf("unexpected response type: %d", resp.Results[0].Type)
-	} else if columns := resp.Results[0].Row.Columns; !reflect.DeepEqual(columns, []uint64{1, SliceWidth + 1}) {
+	} else if columns := resp.Results[0].Row.Columns; !reflect.DeepEqual(columns, []uint64{1, pilosa.SliceWidth + 1}) {
 		t.Fatalf("unexpected columns: %+v", columns)
 	} else if attrs := resp.Results[0].Row.Attrs; len(attrs) != 3 {
 		t.Fatalf("unexpected attr length: %d", len(attrs))
@@ -486,7 +487,7 @@ func TestHandler_Query_Row_ColumnAttrs_Protobuf(t *testing.T) {
 	r.Header.Set("Content-Type", "application/x-protobuf")
 	r.Header.Set("Accept", "application/x-protobuf")
 	h.ServeHTTP(w, r)
-	if w.Code != http.StatusOK {
+	if w.Code != gohttp.StatusOK {
 		t.Fatalf("unexpected status code: %d", w.Code)
 	}
 
@@ -494,9 +495,9 @@ func TestHandler_Query_Row_ColumnAttrs_Protobuf(t *testing.T) {
 	if err := proto.Unmarshal(w.Body.Bytes(), &resp); err != nil {
 		t.Fatal(err)
 	}
-	if columns := resp.Results[0].Row.Columns; !reflect.DeepEqual(columns, []uint64{1, SliceWidth + 1}) {
+	if columns := resp.Results[0].Row.Columns; !reflect.DeepEqual(columns, []uint64{1, pilosa.SliceWidth + 1}) {
 		t.Fatalf("unexpected columns: %+v", columns)
-	} else if rt := resp.Results[0].Type; rt != pilosa.QueryResultTypeRow {
+	} else if rt := resp.Results[0].Type; rt != http.QueryResultTypeRow {
 		t.Fatalf("unexpected response type: %d", resp.Results[0].Type)
 	} else if attrs := resp.Results[0].Row.Attrs; len(attrs) != 3 {
 		t.Fatalf("unexpected attr length: %d", len(attrs))
@@ -536,7 +537,7 @@ func TestHandler_Query_Pairs_JSON(t *testing.T) {
 
 	w := httptest.NewRecorder()
 	h.ServeHTTP(w, test.MustNewHTTPRequest("POST", "/index/i/query", strings.NewReader(`TopN(field=x, n=2)`)))
-	if w.Code != http.StatusOK {
+	if w.Code != gohttp.StatusOK {
 		t.Fatalf("unexpected status code: %d", w.Code)
 	} else if body := w.Body.String(); body != `{"results":[[{"id":1,"count":2},{"id":3,"count":4}]]}`+"\n" {
 		t.Fatalf("unexpected body: %q", body)
@@ -562,14 +563,14 @@ func TestHandler_Query_Pairs_Protobuf(t *testing.T) {
 	r := test.MustNewHTTPRequest("POST", "/index/i/query", strings.NewReader(`TopN(field=x, n=2)`))
 	r.Header.Set("Accept", "application/x-protobuf")
 	h.ServeHTTP(w, r)
-	if w.Code != http.StatusOK {
+	if w.Code != gohttp.StatusOK {
 		t.Fatalf("unexpected status code: %d", w.Code)
 	}
 
 	var resp internal.QueryResponse
 	if err := proto.Unmarshal(w.Body.Bytes(), &resp); err != nil {
 		t.Fatal(err)
-	} else if rt := resp.Results[0].Type; rt != pilosa.QueryResultTypePairs {
+	} else if rt := resp.Results[0].Type; rt != http.QueryResultTypePairs {
 		t.Fatalf("unexpected response type: %d", resp.Results[0].Type)
 	} else if a := resp.Results[0].GetPairs(); len(a) != 2 {
 		t.Fatalf("unexpected pair length: %d", len(a))
@@ -590,7 +591,7 @@ func TestHandler_Query_Err_JSON(t *testing.T) {
 
 	w := httptest.NewRecorder()
 	h.ServeHTTP(w, test.MustNewHTTPRequest("POST", "/index/i/query", strings.NewReader(`Bitmap(id=100)`)))
-	if w.Code != http.StatusBadRequest {
+	if w.Code != gohttp.StatusBadRequest {
 		t.Fatalf("unexpected status code: %d", w.Code)
 	} else if body := w.Body.String(); body != `{"error":"executing: marker"}`+"\n" {
 		t.Fatalf("unexpected body: %q", body)
@@ -613,7 +614,7 @@ func TestHandler_Query_Err_Protobuf(t *testing.T) {
 	r := test.MustNewHTTPRequest("POST", "/index/i/query", strings.NewReader(`TopN(field=x, n=2)`))
 	r.Header.Set("Accept", "application/x-protobuf")
 	h.ServeHTTP(w, r)
-	if w.Code != http.StatusBadRequest {
+	if w.Code != gohttp.StatusBadRequest {
 		t.Fatalf("unexpected status code: %d", w.Code)
 	}
 
@@ -635,7 +636,7 @@ func TestHandler_Query_MethodNotAllowed(t *testing.T) {
 	h.API.Holder = hldr.Holder
 	w := httptest.NewRecorder()
 	h.ServeHTTP(w, test.MustNewHTTPRequest("GET", "/index/i/query", nil))
-	if w.Code != http.StatusMethodNotAllowed {
+	if w.Code != gohttp.StatusMethodNotAllowed {
 		t.Fatalf("invalid status: %d", w.Code)
 	}
 }
@@ -650,7 +651,7 @@ func TestHandler_Query_ErrParse(t *testing.T) {
 	h.API.Holder = hldr.Holder
 	w := httptest.NewRecorder()
 	h.ServeHTTP(w, test.MustNewHTTPRequest("POST", "/index/idx0/query?slices=0,1", strings.NewReader("bad_fn(")))
-	if w.Code != http.StatusBadRequest {
+	if w.Code != gohttp.StatusBadRequest {
 		t.Fatalf("unexpected status code: %d", w.Code)
 	} else if body := w.Body.String(); body != `{"error":"parsing: expected comma, right paren, or identifier, found \"\" occurred at line 1, char 8"}`+"\n" {
 		t.Fatalf("unexpected body: %s", body)
@@ -672,14 +673,14 @@ func TestHandler_Index_Delete(t *testing.T) {
 	}
 
 	// Send request to delete index.
-	resp, err := http.DefaultClient.Do(test.MustNewHTTPRequest("DELETE", s.URL+"/index/i", strings.NewReader("")))
+	resp, err := gohttp.DefaultClient.Do(test.MustNewHTTPRequest("DELETE", s.URL+"/index/i", strings.NewReader("")))
 	if err != nil {
 		t.Fatal(err)
 	}
 	defer resp.Body.Close()
 
 	// Verify body response.
-	if resp.StatusCode != http.StatusOK {
+	if resp.StatusCode != gohttp.StatusOK {
 		t.Fatalf("unexpected status: %d", resp.StatusCode)
 	} else if buf, err := ioutil.ReadAll(resp.Body); err != nil {
 		t.Fatal(err)
@@ -707,7 +708,7 @@ func TestHandler_DeleteField(t *testing.T) {
 	h.API.Cluster = test.NewCluster(1)
 	w := httptest.NewRecorder()
 	h.ServeHTTP(w, test.MustNewHTTPRequest("DELETE", "/index/i0/field/f1", strings.NewReader("")))
-	if w.Code != http.StatusOK {
+	if w.Code != gohttp.StatusOK {
 		t.Fatalf("unexpected status code: %d", w.Code)
 	} else if body := w.Body.String(); body != `{}`+"\n" {
 		t.Fatalf("unexpected body: %s", body)
@@ -749,7 +750,7 @@ func TestHandler_Index_AttrStore_Diff(t *testing.T) {
 	blks[1].Checksum = []byte("MISMATCHED_CHECKSUM")
 
 	// Send block checksums to determine diff.
-	resp, err := http.Post(
+	resp, err := gohttp.Post(
 		s.URL+"/index/i/attr/diff",
 		"application/json",
 		strings.NewReader(`{"blocks":`+string(test.MustMarshalJSON(blks))+`}`),
@@ -799,7 +800,7 @@ func TestHandler_Field_AttrStore_Diff(t *testing.T) {
 	blks[1].Checksum = []byte("MISMATCHED_CHECKSUM")
 
 	// Send block checksums to determine diff.
-	resp, err := http.Post(
+	resp, err := gohttp.Post(
 		s.URL+"/index/i/field/meta/attr/diff",
 		"application/json",
 		strings.NewReader(`{"blocks":`+string(test.MustMarshalJSON(blks))+`}`),
@@ -831,7 +832,7 @@ func TestHandler_Version(t *testing.T) {
 	if strings.HasPrefix(version, "v") {
 		version = version[1:]
 	}
-	if w.Code != http.StatusOK {
+	if w.Code != gohttp.StatusOK {
 		t.Fatalf("unexpected status code: %d", w.Code)
 	} else if w.Body.String() != `{"version":"`+version+`"}`+"\n" {
 		t.Fatalf("unexpected body: %q", w.Body.String())
@@ -851,7 +852,7 @@ func TestHandler_Fragment_Nodes(t *testing.T) {
 	w := httptest.NewRecorder()
 	r := test.MustNewHTTPRequest("GET", "/fragment/nodes?index=X&slice=0", nil)
 	h.ServeHTTP(w, r)
-	if w.Code != http.StatusOK {
+	if w.Code != gohttp.StatusOK {
 		t.Fatalf("unexpected status code: %d", w.Code)
 	} else if body := w.Body.String(); body != `[{"id":"node2","uri":{"scheme":"http","host":"host2"},"isCoordinator":false},{"id":"node0","uri":{"scheme":"http","host":"host0"},"isCoordinator":false}]`+"\n" {
 		t.Fatalf("unexpected body: %q", body)
@@ -861,7 +862,7 @@ func TestHandler_Fragment_Nodes(t *testing.T) {
 	w = httptest.NewRecorder()
 	r = test.MustNewHTTPRequest("GET", "/fragment/nodes?db=X&slice=0", nil)
 	h.ServeHTTP(w, r)
-	if w.Code != http.StatusBadRequest {
+	if w.Code != gohttp.StatusBadRequest {
 		t.Fatalf("unexpected status code: %d", w.Code)
 	}
 
@@ -869,7 +870,7 @@ func TestHandler_Fragment_Nodes(t *testing.T) {
 	w = httptest.NewRecorder()
 	r = test.MustNewHTTPRequest("GET", "/fragment/nodes?slice=0", nil)
 	h.ServeHTTP(w, r)
-	if w.Code != http.StatusBadRequest {
+	if w.Code != gohttp.StatusBadRequest {
 		t.Fatalf("unexpected status code: %d", w.Code)
 	}
 }
@@ -885,7 +886,7 @@ func TestHandler_Expvars(t *testing.T) {
 	w := httptest.NewRecorder()
 	r := test.MustNewHTTPRequest("GET", "/debug/vars", nil)
 	h.ServeHTTP(w, r)
-	if w.Code != http.StatusOK {
+	if w.Code != gohttp.StatusOK {
 		t.Fatalf("unexpected status code: %d", w.Code)
 	}
 }
@@ -908,7 +909,7 @@ func TestHandler_RecalculateCaches(t *testing.T) {
 
 	w := httptest.NewRecorder()
 	h.ServeHTTP(w, test.MustNewHTTPRequest("POST", "/recalculate-caches", nil))
-	if w.Code != http.StatusNoContent {
+	if w.Code != gohttp.StatusNoContent {
 		t.Fatalf("unexpected status code: %d", w.Code)
 	}
 
@@ -939,7 +940,7 @@ func TestHandler_CORS(t *testing.T) {
 	}
 
 	// CORS config should allow preflight response
-	handler = test.MustNewHandler(pilosa.OptHandlerAllowedOrigins([]string{"http://test/"}))
+	handler = test.MustNewHandler(http.OptHandlerAllowedOrigins([]string{"http://test/"}))
 	w = httptest.NewRecorder()
 	handler.ServeHTTP(w, req)
 	result = w.Result()
